Add Windows (MSVC) native build support
Add CMake build system and Windows compatibility layer enabling STAR to
compile and run natively on Windows with Microsoft Visual C++. Verified
byte-identical output to Linux STAR 2.7.11b on 434M-read STARsolo dataset
(cynomolgus macaque, 4626 cells, all alignment stats and count matrices match).

Key changes:
- CMakeLists.txt: cross-platform build with auto-fetched zlib
- wincompat.h: POSIX API shims (pthread, sleep, mkdir, kill, mmap stubs)
- FixedStreamBuf.h: cross-platform streambuf replacing pubsetbuf (no-op on MSVC)
- htslib/win32_compat.h + stub headers for htslib C code
- Replace FIFO+vfork/exec with temp files+system() on Windows
- Replace nftw() with FindFirstFile/RemoveDirectory on Windows
- Replace VLAs with std::vector for C++ standard compliance
- Fix OpenMP loop variables to signed types (MSVC OpenMP 2.0)
- Add missing #include <numeric>

Bug fixes (all platforms):
- Initialize all 11 pthread_mutex_t in ThreadControl (upstream only did 8;
  zero-init works on Linux by accident but crashes on Windows)
- Fix stitchAlignToTranscript.h const mismatch between declaration/definition

Windows limitations:
- Only --genomeLoad NoSharedMemory supported
- --readFilesCommand uses temp files instead of FIFO pipes

+++ source/SoloFeature_countSmartSeq.cpp
@@ -40,7 +40,7 @@ void SoloFeature::countSmartSeq()
     vector<vector<typeFeatureUMI>::iterator> cbFeatUMI (nCB + 1);    
     
     #pragma omp parallel for num_threads(P.runThreadN)
-    for (uint32 ired=0; ired<redistrFilesCBfirst.size()-1; ired++) {//TODO: parallelize, each ired is independent here!
+    for (int64 ired=0; ired<(int64)redistrFilesCBfirst.size()-1; ired++) {//TODO: parallelize, each ired is independent here!
         vector<typeFeatureUMI> vFeatureUMI (redistrFilesNreads[ired]);
         
         uint32 iCB1=redistrFilesCBfirst[ired];
