+++ source/insertSeqSA.cpp
@@ -74,7 +74,7 @@ uint insertSeqSA(PackedArray & SA, PackedArray & SA1, PackedArray & SAi, char * 
 
     #pragma omp parallel num_threads(P.runThreadN)
     #pragma omp for schedule (dynamic,1000)
-    for (uint ii=0; ii<2*nG1; ii++) {//find insertion points for each of the sequences
+    for (int64 ii=0; ii<(int64)(2*nG1); ii++) {//find insertion points for each of the sequences
 
         if (seq1[0][ii]>3)
         {//no index for suffices starting with N
